+++ include/pso.h
@@ -9,11 +9,11 @@ extern "C" {
 /**
    PSO algorithm.
  */
-float *pso_basic(obj_func_t obj_func,
+double *pso_basic(obj_func_t obj_func,
                  size_t swarm_size,
                  size_t dim, size_t max_iter,
-                 const float min_position,
-                 const float max_position);
+                 const double min_position,
+                 const double max_position);
 
 /**
     Initialise an array to random numbers between `min` and `max`.
@@ -24,8 +24,8 @@ float *pso_basic(obj_func_t obj_func,
       min     minimum allowed value of each entry
       max     maximum allowed value of each entry
 */
-void pso_rand_init(float *const array, size_t length,
-                    const float min, const float max);
+void pso_rand_init(double *const array, size_t length,
+                    const double min, const double max);
 
 /**
    Evaluate fitness of `positions` according to `obj_func` and store the result
@@ -40,7 +40,7 @@ void pso_rand_init(float *const array, size_t length,
  */
 void pso_eval_fitness(obj_func_t obj_func,
                       size_t swarm_size, size_t dim,
-                      const float *positions, float *fitness);
+                      const double *positions, double *fitness);
 
 /**
    Randomly generate initial velocity for all particles.
@@ -53,10 +53,10 @@ void pso_eval_fitness(obj_func_t obj_func,
      max_position  upper bound on each entry of the position for a particle
      min_position  lower bound on each entry of the position for a particle
  */
-void pso_gen_init_velocity(float *const velocity, const float *positions,
+void pso_gen_init_velocity(double *const velocity, const double *positions,
                           size_t swarm_size, size_t dim,
-                          const float min_position,
-                          const float max_position);
+                          const double min_position,
+                          const double max_position);
 
 /**
    Returns the index of the particle with lowest fitness.
@@ -68,7 +68,7 @@ void pso_gen_init_velocity(float *const velocity, const float *positions,
    Returns:
      A `size_t` representing the index.
  */
-size_t pso_best_fitness(float *fitness, size_t swarm_size);
+size_t pso_best_fitness(double *fitness, size_t swarm_size);
 
 
 /**
@@ -84,11 +84,11 @@ size_t pso_best_fitness(float *fitness, size_t swarm_size);
      min_vel               minimum allowed value for each velocity entry
      max_vel               maximum allowed value for each velocity entry
  */
-void pso_update_velocity(float *velocity, float *positions,
-                         float *local_best_position,
-                         float *best, size_t swarm_size, size_t dim,
-                         const float min_vel,
-                         const float max_vel);
+void pso_update_velocity(double *velocity, double *positions,
+                         double *local_best_position,
+                         double *best, size_t swarm_size, size_t dim,
+                         const double min_vel,
+                         const double max_vel);
 
 /**
    Update the local best fitness and position for all particles.
@@ -101,8 +101,8 @@ void pso_update_velocity(float *velocity, float *positions,
      swarm_size            number of particles in the swarm]
      dim                   dimension of a single particle
  */
-void pso_update_bests(float *local_best_fitness, float *current_fitness,
-                      float *local_best_positions, float *current_positions,
+void pso_update_bests(double *local_best_fitness, double *current_fitness,
+                      double *local_best_positions, double *current_positions,
                       size_t swarm_size, size_t dim);
 
 /**
@@ -116,10 +116,10 @@ void pso_update_bests(float *local_best_fitness, float *current_fitness,
      min_position  lower bound on each entry of the position for a particle
      max_position  upper bound on each entry of the position for a particle
  */
-void pso_update_position(float *positions, float *velocity,
+void pso_update_position(double *positions, double *velocity,
                          size_t swarm_size, size_t dim,
-                         const float min_position,
-                         const float max_position);
+                         const double min_position,
+                         const double max_position);
 #ifdef __cplusplus
 }
 #endif
