+++ tests/test_hgwosca.c
@@ -11,9 +11,9 @@
 Test(hgwosca_unit, init_population) {
   size_t wolf_count = 40;
   size_t dim = 4;
-  float min = 0.0;
-  float max = 10.0;
-  float population[wolf_count * dim];
+  double min = 0.0;
+  double max = 10.0;
+  double population[wolf_count * dim];
   gwo_init_population(population, wolf_count, dim, min, max);
   for(size_t wolf = 0; wolf < wolf_count * dim; wolf++) {
     cr_expect_leq(population[wolf], max,
@@ -27,13 +27,13 @@ Test(hgwosca_unit, init_population) {
 Test(hgwosca_unit, update_fitness) {
   size_t wolf_count = 4;
   size_t dim = 2;
-  float population[] = {
+  double population[] = {
       0.0, 0.0,
       1.0, 4.0,
       5.0, 0.0,
       2.0, 3.0
   };
-  float fitness[] = {0.0, 0.0, 0.0, 0.0};
+  double fitness[] = {0.0, 0.0, 0.0, 0.0};
   gwo_update_fitness(wolf_count, dim, sum_of_squares, population, fitness);
   cr_expect_float_eq(fitness[0], 0.0, FLT_EPSILON, "first wolf's fitness should be 0");
   cr_expect_float_eq(fitness[1], 17.0, FLT_EPSILON, "second wolf's fitness should be 17");
@@ -45,13 +45,13 @@ Test(hgwosca_unit, update_fitness) {
 Test(hgwosca_unit, init_fitness) {
   size_t wolf_count = 4;
   size_t dim = 2;
-  float population[] = {
+  double population[] = {
       0.0, 0.0,
       1.0, 4.0,
       5.0, 0.0,
       2.0, 3.0
   };
-  float fitness[wolf_count];
+  double fitness[wolf_count];
   gwo_init_fitness(fitness, wolf_count, dim, sum_of_squares, population);
   gwo_update_fitness(wolf_count, dim, sum_of_squares, population, fitness);
   cr_expect_float_eq(fitness[0], 0.0, FLT_EPSILON, "first wolf's fitness should be 0");
@@ -64,7 +64,7 @@ Test(hgwosca_unit, init_fitness) {
 Test(hgwosca_unit, update_leaders) {
   size_t wolf_count = 10;
   size_t alpha = 0, beta = 0, delta = 0;
-  float fitness[] = {1.0, 2.0, 3.0, 3.0, 0.0, 5.0, 2.0, 1.0, 0.0, -1.0};
+  double fitness[] = {1.0, 2.0, 3.0, 3.0, 0.0, 5.0, 2.0, 1.0, 0.0, -1.0};
   gwo_update_leaders(wolf_count, fitness, &alpha, &beta, &delta);
   cr_expect_eq(alpha, 9, "alpha should be 10th wolf");
   cr_expect((beta == 4 && delta == 8) || (delta == 4 && beta == 8),
@@ -72,7 +72,7 @@ Test(hgwosca_unit, update_leaders) {
   fitness[8] = 8.0;
   fitness[4] = -2.0;
   fitness[7] = 5.5;
-  /* float fitness[] = {1.0, 2.0, 3.0, 3.0, -2.0, 5.0, 2.0, 5.5, 8.0, -1.0}; */
+  /* double fitness[] = {1.0, 2.0, 3.0, 3.0, -2.0, 5.0, 2.0, 5.5, 8.0, -1.0}; */
   gwo_update_leaders(wolf_count, fitness, &alpha, &beta, &delta);
   cr_expect_eq(alpha, 4, "alpha should be 5th wolf");
   cr_expect_eq(beta, 9, "beta should be 10th wolf");
@@ -81,11 +81,11 @@ Test(hgwosca_unit, update_leaders) {
 
 
 Test(hgwosca_unit, clamp) {
-  float vals[] = { -10.0, 0.0, 400, -3.0, 5.01};
-  float min = 5.0;
-  float max = 100;
+  double vals[] = { -10.0, 0.0, 400, -3.0, 5.01};
+  double min = 5.0;
+  double max = 100;
   for(size_t idx = 0; idx < 5; idx++) {
-    float res = gwo_clamp(vals[idx], min, max);
+    double res = gwo_clamp(vals[idx], min, max);
     cr_expect_leq(res, max,
                   "clamped value at index %ld is bound above by max", idx);
     cr_expect_geq(res, min,
@@ -95,13 +95,13 @@ Test(hgwosca_unit, clamp) {
 
 
 Test(hgwosca_unit, clamp_all_positions) {
-  float vals[] = {
+  double vals[] = {
       -10.0, 0.0, 400, -3.0, 5.01,
       45, 3, 30, -4, 5,
       -100, -100, -100, -100, 100
   };
-  float min = 5.0;
-  float max = 100;
+  double min = 5.0;
+  double max = 100;
   gwo_clamp_all_positions(3, 5, vals, min, max);
   for(size_t idx = 0; idx < 5; idx++) {
     cr_expect_leq(vals[idx], max,
